+++ src/axios/category.js
@@ -1,7 +1,6 @@
 import request from './axios';
-import { GET_CATEGORY } from '../constants/api';
+import { GET_CATEGORY } from '../constants/API';
 
 export const getCategory = async () => {
   return request.get(GET_CATEGORY());
-  //   return GET_CATEGORY();
 };
